Adding middleware and validations to user authenticated

+++ resources/views/livewire/show-post.blade.php
@@ -14,31 +14,35 @@
                     <p class="text-xs text-gray-500">{{ $post->created_at->diffForHumans() }}</p>
                 </div>
             </div>
-            <div class="">
-                <x-dropdown align="right" width="48">
-                    <x-slot name="trigger">
-                        <button class="inline-flex items-center px-3 py-2 border border-transparent text-sm leading-4 font-medium rounded-md text-gray-500 bg-white hover:text-gray-700 focus:outline-none transition ease-in-out duration-150">
-                            <div class="ms-1">
-                                <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
-                                    <path stroke-linecap="round" stroke-linejoin="round" d="M12 6.75a.75.75 0 1 1 0-1.5.75.75 0 0 1 0 1.5ZM12 12.75a.75.75 0 1 1 0-1.5.75.75 0 0 1 0 1.5ZM12 18.75a.75.75 0 1 1 0-1.5.75.75 0 0 1 0 1.5Z" />
-                                </svg>
-                            </div>
-                        </button>
-                    </x-slot>
+            @auth
+                @if($post->user_id === auth()->user()->id)
+                    <div>
+                        <x-dropdown align="right" width="48">
+                            <x-slot name="trigger">
+                                <button class="inline-flex items-center px-3 py-2 border border-transparent text-sm leading-4 font-medium rounded-md text-gray-500 bg-white hover:text-gray-700 focus:outline-none transition ease-in-out duration-150">
+                                    <div class="ms-1">
+                                        <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-6 h-6">
+                                            <path stroke-linecap="round" stroke-linejoin="round" d="M12 6.75a.75.75 0 1 1 0-1.5.75.75 0 0 1 0 1.5ZM12 12.75a.75.75 0 1 1 0-1.5.75.75 0 0 1 0 1.5ZM12 18.75a.75.75 0 1 1 0-1.5.75.75 0 0 1 0 1.5Z" />
+                                        </svg>
+                                    </div>
+                                </button>
+                            </x-slot>
 
-                    <x-slot name="content">
-                        <x-dropdown-link :href="route('posts.edit', $post)">
-                            {{ __('Edit') }}
-                        </x-dropdown-link>
-                        <button wire:click="deletePost({{ $post->id }})"
-                                wire:confirm="Are you sure you want to delete this post?"
-                                class="block w-full px-4 py-2 text-start text-sm leading-5 hover:bg-indigo-50 focus:outline-none focus:bg-gray-50 transition duration-150 ease-in-out"
-                        >
-                            <p class="text-gray-700 hover:text-indigo-600 hover:underline">Delete</p>
-                        </button>
-                    </x-slot>
-                </x-dropdown>
-            </div>
+                            <x-slot name="content">
+                                <x-dropdown-link :href="route('posts.edit', $post)">
+                                    {{ __('Edit') }}
+                                </x-dropdown-link>
+                                <button wire:click="deletePost({{ $post->id }})"
+                                        wire:confirm="Are you sure you want to delete this post?"
+                                        class="block w-full px-4 py-2 text-start text-sm leading-5 hover:bg-indigo-50 focus:outline-none focus:bg-gray-50 transition duration-150 ease-in-out"
+                                >
+                                    <p class="text-gray-700 hover:text-indigo-600 hover:underline">Delete</p>
+                                </button>
+                            </x-slot>
+                        </x-dropdown>
+                    </div>
+                @endif
+            @endauth
         </div>
         <p class="text-xl md:text-4xl font-extrabold">{{ $post->title }}</p>
         <p class="text-md md:text-xl">{{ $post->content }}</p>
@@ -91,9 +95,9 @@
                     </div>
                     <div class="w-full border border-gray-300 rounded-lg px-3 py-2">
                         <div class="flex gap-2 items-center">
-                            <p class="text-md text-gray-700 font-bold">{{ $post->user->name }}</p>
+                            <p class="text-md text-gray-700 font-bold">{{ $comment->user->name}}</p>
                             <p class="font-bold text-2xl text-gray-500">·</p>
-                            <p class="text-xs text-gray-500">{{ $post->created_at->diffForHumans() }}</p>
+                            <p class="text-xs text-gray-500">{{ $comment->created_at->diffForHumans() }}</p>
                         </div>
                         <p class="mt-1">{{ $comment->comment }}</p>
                     </div>
